refactor(settings): move form components into settings/ subdirectory
Rename from settings-advanced-*.tsx to settings/*.tsx for cleaner
file paths without redundant "advanced" prefix.
