+++ src/templates/admin/settings/email-tpl-admin.tsx
