+++ src/templates/admin/settings/email-tpl-confirmation.tsx
