+++ src/templates/admin/settings/email.tsx
