+++ src/templates/admin/settings/google-wallet.tsx
